services/api-service/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database — async SQLAlchemy
sqlalchemy==2.0.36
aiosqlite==0.20.0       # SQLite driver (local/test)
asyncpg==0.30.0         # PostgreSQL driver (staging/prod)

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
email-validator==2.2.0

# AWS SDK
boto3==1.36.14

# Structured logging
python-json-logger==3.2.1

# OpenTelemetry — Tracing
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0

# OpenTelemetry — Metrics
opentelemetry-exporter-otlp-proto-http==1.29.0

# Health check support
httpx==0.28.1

# Prometheus metrics
prometheus-fastapi-instrumentator==7.0.0

# Pin sse-starlette to a version compatible with starlette 0.41.x
sse-starlette==1.6.5
